Fix infinite loop issue freezing the game
An image created from `toDataURL` sometimes has a `width` and
height of `0` for a brief moment. If this happens, it can cause
the game to get stuck in an infinite loop and freeze, because
the draw loop tries to shift each draw to the right/down by
`0` pixels, which goes nowhere, so the loop can never terminate.